+++ src/server/dashboard/mod.rs
@@ -11,13 +11,11 @@ impl Plugin for DashboardPlugin {
     fn build(&self, app: &mut App) {
         info!("Building DashboardPlugin");
 
-        // Register events
         app.add_event::<events::ExampleEvent>();
 
-        // Add systems
-        app.add_systems(Update, systems::example_system);
+        app.add_systems(Update, (systems::run_basic_ui));
+        app.add_systems(Update, (systems::quit_system));
 
-        // Add resources
         app.insert_resource(resources::ExampleResource::default());
-      }
     }
+}
